requirements.txt
torch>=2.0.1
transformers>=4.30.2
opencv-python-headless>=4.7.0
yt-dlp>=2023.7.6
openai-whisper
fastapi>=0.95.2
uvicorn>=0.22.0
redis>=4.5.5
python-multipart>=0.0.6
clip@git+https://github.com/openai/CLIP.git
ultralytics>=8.0.0
numpy>=1.24.3
better-profanity>=0.7.0
nltk>=3.8.2